Refactor again
- move pipeline files to lib/
