--- azure/pipeline/lib/README.md
@@ -1,5 +1,0 @@
-# Pipeline Lib
-
-Files needed to run the pipeline in `azure/pipeline`. 
-
-They are defined here as I want to consolidate all "imported" files in one place.
