--- azure/pipeline/lib/__init__.py
@@ -1,3 +1,0 @@
-# an empty __init__.py
-# its existence is required for the files in this folder to be importable to other files
-# such as `azure/deploy/main.py` and `tests/endpoint/main.py`
